Refactor ShowDetailScreen into component-based architecture
