+++ src/screens/ShowDetail/components/index.ts
@@ -1,0 +1,17 @@
+import ShowHeaderActions from './ShowHeaderActions';
+import ShowBasicInfo from './ShowBasicInfo';
+import ShowTimeInfo from './ShowTimeInfo';
+import OrganizerInfo from './OrganizerInfo';
+import DealersList from './DealersList';
+import ShowDescription from './ShowDescription';
+import ShowManagementButtons from './ShowManagementButtons';
+
+export {
+  ShowHeaderActions,
+  ShowBasicInfo,
+  ShowTimeInfo,
+  OrganizerInfo,
+  DealersList,
+  ShowDescription,
+  ShowManagementButtons
+};
